refactor(termgfx): the geometry of an image and the cursor contract move out of ic
