+++ src/ic/Cargo.toml
@@ -18,7 +18,6 @@ image.workspace = true
 notify.workspace = true
 termion.workspace = true
 termgfx.workspace = true
-termsize.workspace = true
 unicode-width.workspace = true
 
 [lints]
